Refactor last page handling
In pagination functionality, the last page number was used to determine whether or not it was possible to fetch the next page. This was replaced with a built-in boolean flag provided by the backend.

Also, previous page elements in petsInfo were discarded. Now they are kept in the new state with the newly fetched elements.